requirements.txt
Django==4.0.5
djangorestframework==3.13.1
django-cors-headers==3.13.0
https://github.com/EgorGurtsiev/djangorestframework-simplejwt.git
requests==2.28.1
